+++ aiobungie/url.py
@@ -24,7 +24,7 @@
 
 from __future__ import annotations
 
-__all__ = ("BASE", "REST_EP", "OAUTH_EP", "TOKEN_EP", "OAUTH2_EP_BUILDER")
+__all__ = ("BASE", "REST_EP", "OAUTH_EP", "TOKEN_EP")
 
 import typing
 
@@ -39,22 +39,3 @@ OAUTH_EP: typing.Final[str] = f"{BASE}/en/OAuth/Authorize"
 
 TOKEN_EP: typing.Final[str] = "/App/OAuth/token"
 """OAuth token endpoint"""
-
-OAUTH2_EP_BUILDER: typing.Final[
-    str
-] = "{oauth_endpoint}?client_id={client_id}&response_type=code&state={uuid}"
-"""Builds an OAuth2 authorize URL given an application client id.
-
-Parameters may be passed as kwargs using `str.format` method. i.e.,
-
-Example
--------
-```py
-import aiobungie
-import uuid
-
-aiobungie.url.OAUTH2_EP_BUILDER.format(
-    oauth_endpoint=aiobungie.url.OAUTH_EP, client_id=1234, uuid=str(uuid.uuid4())
-)
-```
-"""